refactor: 디바이스 토큰 vo 경로 변경 (#236)
