+++ infrastructure/mysql/src/main/kotlin/kr/wooco/woocobe/mysql/notification/DeviceTokenPersistenceAdapter.kt
@@ -3,9 +3,9 @@ package kr.wooco.woocobe.mysql.notification
 import kr.wooco.woocobe.core.notification.application.port.out.DeviceTokenCommandPort
 import kr.wooco.woocobe.core.notification.application.port.out.DeviceTokenQueryPort
 import kr.wooco.woocobe.core.notification.domain.entity.DeviceToken
+import kr.wooco.woocobe.core.notification.domain.entity.DeviceToken.Token
 import kr.wooco.woocobe.core.notification.domain.exception.NotExistsDeviceTokenException
 import kr.wooco.woocobe.core.notification.domain.vo.DeviceTokenStatus
-import kr.wooco.woocobe.core.notification.domain.vo.Token
 import kr.wooco.woocobe.mysql.notification.repository.DeviceTokenJpaRepository
 import org.springframework.stereotype.Component
